smb: move common parsing funcs into own file
